requirements.txt
albumentations==0.5.2
mayavi==4.8.1
MedPy==0.4.0
numpy==1.18.5
opencv_python==4.2.0.32
Pillow==9.5.0
scikit_image==0.19.1
scikit_learn==1.2.2
scipy==1.4.1
SimpleITK==2.2.1
skimage==0.0
torch==1.8.0
torchio==0.18.53
torchvision==0.9.0
visdom==0.1.8.9
